backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic-settings
sqlalchemy
psycopg[binary]
langchain
langchain-text-splitters
langchain-openai
openai
pypdf2
